Quick fix for to add float numbers

+++ Data/Calculator(general).py
@@ -6,23 +6,23 @@ opperation2 = opperation.lower()
 if opperation2.__contains__("add"):
     number1 = input("What number would you like to add?")
     number2 = input("What number would you like to add to the first number?")
-    total = int(number1) + int(number2)
+    total = float(number1) + float(number2)
     print(total)
 
 if opperation2.__contains__("subtract"):
     number1 = input("What number would you like to subtract from?")
     number2 = input("What number would you like to subtract from the first number?")
-    total = int(number1) - int(number2)
+    total = float(number1) - float(number2)
     print(total)
 
 if opperation2.__contains__("multiply"):
     number1 = input("What number would you like to multiply?")
     number2 = input("What number would you like to multiply the first number by?")
-    total = int(number1) * int(number2)
+    total = float(number1) * float(number2)
     print(total)
 
 if opperation2.__contains__("divide"):
     number1 = input("What number would you like to divide?")
     number2 = input("What number would you like to divide the first number by?")
-    total = int(number1) / int(number2)
+    total = float(number1) / float(number2)
     print(total)
